improve script and headers

+++ update.py
@@ -6,13 +6,32 @@ import os
 import time
 from datetime import datetime
 
+requests.packages.urllib3.disable_warnings()
+
 portaleURL = "https://www.portaleacque.salute.gov.it/PortaleAcquePubblico/"
 comuniURL = "https://raw.githubusercontent.com/opendatasicilia/comuni-italiani/main/dati/comuni.csv"
 con = sqlite3.connect('./data/latest.db')
 cur = con.cursor()
 req = requests.Session()
-req.headers.update(
-    {'User-Agent': 'Mozilla/5.0 (Windows NT 6.1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.72 Safari/537.36'})
+
+headers = {
+    "Connection": "keep-alive",
+    "sec-ch-ua": '"Not A;Brand";v="99", "Chromium";v="99", "Google Chrome";v="99"',
+    "Accept": "application/json",
+    'Content-Type': 'application/json',
+    "sec-ch-ua-mobile": "?0",
+    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/99.0.4844.74 Safari/537.36",
+    "sec-ch-ua-platform": '"Windows"',
+    "Origin": "https://www.portaleacque.salute.gov.it/",
+    'Sec-Fetch-Site': 'same-origin',
+    'Sec-Fetch-Mode': 'cors',
+    'Sec-Fetch-Dest': 'empty',
+    'Referer': 'https://www.portaleacque.salute.gov.it/PortaleAcquePubblico/',
+    'Accept-Language': 'en-US,en;q=0.9,it;q=0.8',
+    'Cookie': 'cookies_consent=true'
+}
+
+req.headers.update(headers)
 
 cur.execute('''
 create table if not exists areas (
@@ -66,13 +85,15 @@ codici_istat = {
 }
 
 comuni = pd.read_csv(comuniURL)
-
 output = []
 
 
 def downJoin(istat):
     global output
-    data = req.get(portaleURL+'rest/layer/AB/'+istat)
+    try:
+        data = req.get(portaleURL+'rest/layer/AB/'+istat, verify=False)
+    except Exception as e:
+        print(e)
     if data.status_code != 200:
         print(data.status_code)
         return
@@ -87,7 +108,6 @@ def downJoin(istat):
             },
             "geometry": geometry
         }
-        # print(codice)
         output.append(converted)
 
 
@@ -103,14 +123,15 @@ def getRegione(provincia):
 
 def insert(featureCollection):
     # https://www.portaleacque.salute.gov.it/PortaleAcquePubblico/datiArea.do?codiceArea={codice}&tipoArea=undefined&isFuoriNorma=undefined
+    timeout = 0
     areas = []
     missing = []
     cur.execute('BEGIN TRANSACTION;')
-    for record in featureCollection['features']:
-        time.sleep(1)
+    for index, record in enumerate(featureCollection['features']):
+        time.sleep(timeout)
         area = record['properties']['CODICE']
         data = req.get(portaleURL+'datiArea.do?codiceArea=' +
-                       str(area)+'&tipoArea=undefined&isFuoriNorma=undefined')
+                       str(area)+'&tipoArea=undefined&isFuoriNorma=undefined', verify=False)
         if data.status_code == 200:
             data = data.json()
             converted = {
@@ -137,7 +158,7 @@ def insert(featureCollection):
             }
             try:
                 cur.execute("INSERT OR REPLACE INTO areas VALUES (:CODICE, :nome, :comune, :provincia, :siglaProvincia, :regione, :stato, :limiteEi, :limiteEc, :dataInizioStagioneBalneare, :dataFineStagioneBalneare, :statoDesc, :geometry, :isFuoriNorma, :ultimaAnalisi, :valoreEi, :valoreEc, :flagOltreLimiti, :scheda, :interdizioni);", converted)
-                print(converted['CODICE'], converted['nome'], converted['comune'],
+                print('['+str(index)+'/'+str(len(featureCollection['features']))+']', converted['CODICE'], converted['nome'], converted['comune'],
                       converted['provincia'], converted['regione'])
                 areas.append(converted)
             except Exception as e:
@@ -152,23 +173,24 @@ def insert(featureCollection):
     with open('./data/'+timestamp+'.json', 'w+') as f:
         json.dump(areas, f)
     os.rename('./data/latest.db', './data/'+timestamp+'.db')
-    os.remove('./data/output.json')
     print('missing: ', missing)
 
 
 def execute():
-    global output
-    for istat in codici_istat.values():
-        downJoin(istat)
-    print('Looking for duplicates...')
-    nodups = {i['properties']['CODICE']: i for i in reversed(output)}.values()
-    features = list(nodups)
-    output = {"type": "FeatureCollection", "features": features}
-    print('Saving file...')
-    with open('./data/output.json', 'w+') as f:
-        json.dump(output, f)
-    os.system(
-        'mapshaper -i ./data/output.json -snap -simplify weighted 12% keep-shapes -o ./data/minified.json')
+    if not os.path.exists('./data/minified.json'):
+        global output
+        for istat in codici_istat.values():
+            downJoin(istat)
+        print('Looking for duplicates...')
+        nodups = {i['properties']['CODICE']: i for i in reversed(output)}.values()
+        features = list(nodups)
+        output = {"type": "FeatureCollection", "features": features}
+        print('Saving file...')
+        with open('./data/output.json', 'w+') as f:
+            json.dump(output, f)
+        os.system(
+            'mapshaper -i ./data/output.json -snap -simplify weighted 12% keep-shapes -o ./data/minified.json')
+        os.remove('./data/output.json')
     minified = open('./data/minified.json')
     featureCollection = json.loads(minified.read())
     print('Inserting features into db...')
